combine python scripts
