+++ mako-templates/branches.html
@@ -1,8 +1,6 @@
 <%inherit file="base.html"/>
 <%namespace file="base.html" import="*" />
 
-<h1>Branches</h1>
-
 <p>Suppose you are working on a calculator program.
 You start working on a new multiplication feature,
 and you have already <a href="committing.html">committed and pushed</a> several times.
