+++ mako-templates/committing.html
@@ -1,8 +1,6 @@
 <%inherit file="base.html"/>
 <%namespace file="base.html" import="*" />
 
-<h1>Committing</h1>
-
 <p>A commit is a pile of changes. For example
 <a href="https://github.com/Akuli/git-guide/commit/07ee936f719060550b2033ac6501b31d52b8f627">this commit</a>
 changed two files in this guide:
