+++ mako-templates/getting-started.html
@@ -1,8 +1,6 @@
 <%inherit file="base.html"/>
 <%namespace file="base.html" import="*" />
 
-<h1>Getting started with Git and GitHub</h1>
-
 <p>This page contains everything you need to know for getting started with GitHub.
