--- pagelist.py
@@ -1,5 +1,0 @@
-pagelist = [
-    ("getting-started.html", "Getting started", "installing Git, making and cloning a repo"),
-    ("committing.html", "Committing", "add, commit, push, status, diff"),
-    ("branches.html", "Branches", "checkout, log, merge"),
-]
